+++ src/main/java/storage/Decoder.java
@@ -1,0 +1,4 @@
+package storage;
+
+public class Decoder {
+}
