+++ src/main/java/taskclasses/Deadline.java
@@ -1,0 +1,8 @@
+package taskclasses;
+
+public class Deadline extends Task {
+    public Deadline(String description, String by){
+        super(description, "D");
+        Time = by;
+    }
+}
