+++ src/main/java/taskclasses/Todo.java
@@ -1,0 +1,7 @@
+package taskclasses;
+
+public class Todo extends Task{
+    public Todo(String description){
+        super(description, "T");
+    }
+}
